+++ glimpse/ui-compose/src/commonMain/kotlin/GlimpseViewScope.kt
@@ -14,22 +14,15 @@
  * limitations under the License.
  */
 
-package graphics.glimpse.examples.triangle
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import graphics.glimpse.ui.compose.GlimpseView
+package graphics.glimpse.ui.compose
 
 /**
- * Composable application content layout.
+ * Receiver scope used by [GlimpseView].
  */
-@Composable
-fun AppContent(modifier: Modifier = Modifier, resources: AppResources) {
-    GlimpseView(
-        callback = TriangleCallback(resources),
-        modifier = Modifier.fillMaxSize().then(modifier)
-    ) {
-        zOrderOnTop = true
-    }
+expect class GlimpseViewScope {
+
+    /**
+     * Controls whether the surface of the [GlimpseView] is placed on top of its window.
+     */
+    var zOrderOnTop: Boolean
 }